backend/requirements.txt
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.1
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.105.0
fastapi-mail==1.4.1
h11==0.14.0
idna==3.6
imagekitio==3.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openpyxl==3.1.2
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pymongo==4.6.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
requests==2.31.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
urllib3==1.26.18
uvicorn==0.24.0.post1
